trajectory/requirements.txt
pyserial
threaded